document_rag_project/document_rag_project/api/views/baseapiview.py
from rest_framework.views import APIView
from rest_framework import status
from rest_framework.response import Response
from django.conf import Settings
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM
import torch

class BaseAPIView(APIView):
    pass