document_rag_project/document_rag_project/requirements.txt
django
djangorestframework
mysqlclient
transformers
torch